+++ api/src/main/java/net/jqwik/api/lifecycle/AfterExample.java
@@ -15,7 +15,7 @@ import static org.apiguardian.api.API.Status.*;
 @Target({ElementType.ANNOTATION_TYPE, ElementType.METHOD})
 @Retention(RetentionPolicy.RUNTIME)
 @AfterProperty
-@API(status = EXPERIMENTAL, since = "1.2.5")
+@API(status = MAINTAINED, since = "1.4.0")
 public @interface AfterExample {
 
 }