+++ api/src/main/java/net/jqwik/api/lifecycle/CannotResolveParameterException.java
@@ -11,13 +11,13 @@ import static org.apiguardian.api.API.Status.*;
 /**
  * Experimental feature. Not ready for public usage yet.
  */
-@API(status = EXPERIMENTAL, since = "1.2.4")
+@API(status = MAINTAINED, since = "1.4.0")
 public class CannotResolveParameterException extends JqwikException {
 	public CannotResolveParameterException(ParameterResolutionContext context, String info) {
 		this(context.parameter(), info);
 	}
 
-	@API(status = EXPERIMENTAL, since = "1.2.5")
+	@API(status = INTERNAL)
 	public CannotResolveParameterException(Parameter parameter, String info) {
 		super(createMessage(parameter, info));
 	}