+++ api/src/main/java/net/jqwik/api/lifecycle/ContainerLifecycleContext.java
@@ -7,7 +7,7 @@ import static org.apiguardian.api.API.Status.*;
 /**
  * The context of a test container (a container class or the whole jqwik suite).
  */
-@API(status = EXPERIMENTAL, since = "1.0")
+@API(status = MAINTAINED, since = "1.4.0")
 public interface ContainerLifecycleContext extends LifecycleContext {
 
 }