+++ api/src/main/java/net/jqwik/api/lifecycle/LifecycleHooks.java
@@ -11,7 +11,7 @@ import static org.apiguardian.api.API.Status.*;
  */
 @Target({ ElementType.ANNOTATION_TYPE, ElementType.METHOD, ElementType.TYPE })
 @Retention(RetentionPolicy.RUNTIME)
-@API(status = EXPERIMENTAL, since = "1.0")
+@API(status = MAINTAINED, since = "1.4.0")
 public @interface LifecycleHooks {
 	AddLifecycleHook[] value();
 }