+++ api/src/main/java/net/jqwik/api/lifecycle/ParameterResolutionContext.java
@@ -11,7 +11,7 @@ import static org.apiguardian.api.API.Status.*;
 /**
  * Experimental feature. Not ready for public usage yet.
  */
-@API(status = EXPERIMENTAL, since = "1.2.4")
+@API(status = MAINTAINED, since = "1.4.0")
 public interface ParameterResolutionContext {
 
 	Parameter parameter();