+++ api/src/main/java/net/jqwik/api/lifecycle/PropagationMode.java
@@ -14,7 +14,7 @@ import static org.apiguardian.api.API.Status.*;
  *     <li>Use annotation attribute {@linkplain AddLifecycleHook#propagateTo()}</li>
  * </ul>
  */
-@API(status = EXPERIMENTAL, since = "1.2.4")
+@API(status = MAINTAINED, since = "1.4.0")
 public enum PropagationMode {
 
 	/**