+++ api/src/main/java/net/jqwik/api/lifecycle/PropertyLifecycleContext.java
@@ -9,7 +9,7 @@ import static org.apiguardian.api.API.Status.*;
 /**
  * The context information of a property or example.
  */
-@API(status = EXPERIMENTAL, since = "1.0")
+@API(status = MAINTAINED, since = "1.4.0")
 public interface PropertyLifecycleContext extends LifecycleContext {
 
 	/**